Fix status_update()
In master it is already working but not in moo. Also I'm setting the type to 0 because the variable is named game (easier to set in the client when you can just omit it). Now with the rich presence stuff there is way more possible now (activity-object). I might work on that later.