on_finish gets $tx from $self
For some reason I was passing $tx as an argument, and in one case was
passing "-1" if there was no transaction object yet. This was causing
the on_finish sub to fail. Now it just gets $tx from $self and will pass
the validation logic correctly.

I also tried to condense that logic a bit.