Rewrite of connection code
- More focus on validating the connection status before sending anything
- Better logging of connection status and failures
scale back later)
- $ua is built with a lazy loader instead of sub BUILD
- No more manually calling on_finish.
- Better heartbeat logic
- Removed some unnecessary Data::Dumper log entries
- Simplified calls to gw_connect, now takes a hash with the 'resume'
attribute inside, doesn't need a URL passed to it.
- Reconnect Timer Backoff logic should actually work now - doubles the
time between each reconnect (Currently does not reset)

The goal here is to improve the bot's ability to recover from a
disconnect, to know when it can and can't recover automatically and
handle both scenarios appropariately instead of just spamming failed
connection attempts forever until the token gets disabled.

More work still needs to be done particularly in the "what if we can't
reconnect?" area, but this is a start.