reconnect backoff timer reset logic, fix RESUME
- dispatch_ready was not capturing the session ID, so our RESUME packets
on reconnect were being rejected (for not having a session ID). Fixed.

- Reconnect timer was never resetting, so every reconnect doubled the
time between connection attempts. After a couple days like that you
naturally end up with a reconnect timer of 10-20 minutes or more.
- New logic tracks the bot's uptime and if the bot had been connected
for at least 60 seconds before it lost connection, the timer resets.
This way if the bot is disconnecting immediately, the timer will
continue to back off. But if it stays connected for a while, the next
reconnect should be considered "unrelated" and it is probably safe to
reconnect right away.